requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
astor==0.8.1
async-timeout==4.0.3
attrs==23.2.0
bandit==1.7.9
coverage==7.6.0
darglint==1.8.1
docutils==0.21.2
environs==11.0.0
eradicate==2.3.0
exceptiongroup==1.2.2
flake8==7.1.0
flake8-bandit==4.1.1
flake8-broken-line==1.0.0
flake8-bugbear==24.4.26
flake8-commas==2.1.0
flake8-comprehensions==3.15.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-isort==6.1.1
flake8-quotes==3.4.0
flake8-rst-docstrings==0.3.0
flake8-string-format==0.3.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
iniconfig==2.0.0
isort==5.13.2
markdown-it-py==3.0.0
marshmallow==3.21.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
packaging==24.1
pbr==6.0.0
pep8-naming==0.13.3
pluggy==1.5.0
pycodestyle==2.12.0
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
python-dotenv==1.0.1
PyYAML==6.0.1
restructuredtext-lint==1.4.0
rich==13.7.1
smmap==5.0.1
snowballstemmer==2.2.0
stevedore==5.2.0
tomli==2.0.1
typing_extensions==4.12.2
wemake-python-styleguide==0.19.2
yarl==1.9.4
